add api, start splitting the functionality
